(PC-18099)[PRO] feat: Created hook for 500 biggest offerers remote config.
