+++ pro/src/components/pages/Home/Homepage.tsx
@@ -1,11 +1,10 @@
-import { getValue } from '@firebase/remote-config'
-import React, { useEffect, useRef, useState } from 'react'
+import React, { useRef, useState } from 'react'
 
 import { api } from 'apiClient/api'
 import PageTitle from 'components/layout/PageTitle/PageTitle'
 import useActiveFeature from 'hooks/useActiveFeature'
 import useCurrentUser from 'hooks/useCurrentUser'
-import useRemoteConfig from 'hooks/useRemoteConfig'
+import useStats from 'hooks/useStats'
 import { BannerRGS } from 'new_components/Banner'
 import { Newsletter } from 'new_components/Newsletter'
 
@@ -23,31 +22,13 @@ const Homepage = (): JSX.Element => {
   const [hasClosedRGSBanner, setHasClosedRGSBanner] = useState<boolean>(
     Boolean(hasSeenProRgs)
   )
-  const [canSeeStats, setCanSeeStats] = useState<boolean>(false)
+  const canSeeStats = useStats()
   const isOffererStatsActive = useActiveFeature('ENABLE_OFFERER_STATS')
   const handleCloseRGSBanner = () => {
     api.patchProUserRgsSeen().finally(() => {
       setHasClosedRGSBanner(true)
     })
   }
-  const { remoteConfig } = useRemoteConfig()
-  useEffect(() => {
-    if (isOffererStatsActive) {
-      api.listOfferersNames().then(receivedOffererNames => {
-        let biggest500: string[] = []
-        if (remoteConfig) {
-          biggest500 = getValue(remoteConfig, 'only500BiggerActors')
-            .asString()
-            .split(',')
-        }
-        setCanSeeStats(
-          receivedOffererNames.offerersNames.some(v =>
-            biggest500.includes(v.id)
-          )
-        )
-      })
-    }
-  }, [remoteConfig])
 
   return (
     <div className="homepage">
